Average Price Data Grid - ok

+++ IPG_zad2/Model/LevelAveragePrice.cs
@@ -9,6 +9,6 @@ namespace IPG_zad2.Model
     public class LevelAveragePrice
     {
         public string Level { get; set; }
-        public double AveragePrice { get; set; }
+        public float AveragePrice { get; set; }
     }
 }
